# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ClassMaterial": {
				"properties": {
					"file_type": {
						"type": "string",
						"nullable": true
					},
					"file_url": {
						"type": "string",
						"nullable": true
					},
					"display_name": {
						"type": "string"
					},
					"file_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"file_type",
					"file_url",
					"display_name",
					"file_id"
				],
				"type": "object"
			},
			"CourseClassGetResponse": {
				"properties": {
					"materials": {
						"items": {
							"$ref": "#/components/schemas/ClassMaterial"
						},
						"type": "array"
					},
					"fk_module": {
						"type": "number",
						"format": "double"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"class_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"materials",
					"fk_module",
					"index_order",
					"description",
					"title",
					"class_id"
				],
				"type": "object"
			},
			"DefaultSelection_Prisma._36_Course_classPayload_": {
				"properties": {
					"fk_module": {
						"type": "number",
						"format": "double"
					},
					"class_id": {
						"type": "number",
						"format": "double"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"fk_module",
					"class_id",
					"index_order",
					"description",
					"title"
				],
				"type": "object"
			},
			"Course_class": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_Course_classPayload_",
				"description": "Model Course_class"
			},
			"CourseClassUpdateResponse": {
				"properties": {
					"class": {
						"$ref": "#/components/schemas/Course_class"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"class",
					"message"
				],
				"type": "object"
			},
			"Partial_CourseClassCreateRequest_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"index_order": {
						"type": "number",
						"format": "double"
					},
					"fk_module": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CourseClassUpdateRequest": {
				"$ref": "#/components/schemas/Partial_CourseClassCreateRequest_"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ribbit_api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/ribbit/courses": {
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"courses": {
											"items": {
												"properties": {
													"slug": {
														"type": "string"
													},
													"banner_url": {
														"type": "string"
													},
													"description": {
														"type": "string"
													},
													"title": {
														"type": "string"
													},
													"id_course": {
														"type": "number",
														"format": "double"
													},
													"teacher": {
														"properties": {
															"avatar_url": {
																"type": "string"
															},
															"full_name": {
																"type": "string"
															}
														},
														"required": [
															"avatar_url",
															"full_name"
														],
														"type": "object"
													}
												},
												"required": [
													"slug",
													"banner_url",
													"description",
													"title",
													"id_course",
													"teacher"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total_courses": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"courses",
										"total_courses"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCourse",
				"responses": {
					"201": {
						"description": "Criado",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"course_id": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"course_id",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Slug já em uso"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"slug": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"banner": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"title",
									"slug"
								]
							}
						}
					}
				}
			}
		},
		"/ribbit/courses/user/{userId}": {
			"get": {
				"operationId": "FindByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"courses": {
											"items": {
												"properties": {
													"slug": {
														"type": "string"
													},
													"banner_url": {
														"type": "string"
													},
													"description": {
														"type": "string"
													},
													"title": {
														"type": "string"
													},
													"id_course": {
														"type": "number",
														"format": "double"
													},
													"teacher": {
														"properties": {
															"avatar_url": {
																"type": "string"
															},
															"full_name": {
																"type": "string"
															}
														},
														"required": [
															"avatar_url",
															"full_name"
														],
														"type": "object"
													}
												},
												"required": [
													"slug",
													"banner_url",
													"description",
													"title",
													"id_course",
													"teacher"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total_courses": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"courses",
										"total_courses"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ribbit/courses/{courseId}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"slug": {
											"type": "string"
										},
										"banner_url": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id_course": {
											"type": "number",
											"format": "double"
										},
										"modules": {
											"items": {
												"properties": {
													"classes": {
														"items": {
															"properties": {
																"is_completed": {
																	"type": "boolean"
																},
																"title": {
																	"type": "string"
																},
																"class_id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"is_completed",
																"title",
																"class_id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"index_order": {
														"type": "number",
														"format": "double"
													},
													"title": {
														"type": "string"
													},
													"module_id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"classes",
													"index_order",
													"title",
													"module_id"
												],
												"type": "object"
											},
											"type": "array"
										},
										"progress": {
											"type": "number",
											"format": "double"
										},
										"teacher_name": {
											"type": "string"
										}
									},
									"required": [
										"slug",
										"banner_url",
										"description",
										"title",
										"id_course",
										"modules",
										"progress",
										"teacher_name"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Curso não encontrado"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCourse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"course": {
											"properties": {
												"slug": {
													"type": "string"
												},
												"banner_url": {
													"type": "string"
												},
												"description": {
													"type": "string"
												},
												"title": {
													"type": "string"
												},
												"id_course": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"slug",
												"banner_url",
												"description",
												"title",
												"id_course"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"course",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Curso não encontrado"
					},
					"409": {
						"description": "Slug já em uso"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"slug": {
										"type": "string"
									},
									"banner": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteById",
				"responses": {
					"204": {
						"description": "Deletado com sucesso"
					},
					"404": {
						"description": "Curso não encontrado"
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/ribbit/enrollments/my-courses": {
			"get": {
				"operationId": "FindByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"courses": {
											"items": {
												"properties": {
													"enrollment_id": {
														"type": "number",
														"format": "double"
													},
													"enrolled_at": {
														"type": "string",
														"format": "date-time"
													},
													"progress": {
														"type": "number",
														"format": "double"
													},
													"course": {
														"properties": {
															"slug": {
																"type": "string"
															},
															"banner_url": {
																"type": "string"
															},
															"title": {
																"type": "string"
															},
															"course_id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"slug",
															"banner_url",
															"title",
															"course_id"
														],
														"type": "object"
													}
												},
												"required": [
													"enrollment_id",
													"enrolled_at",
													"progress",
													"course"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total_enrollments": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"courses",
										"total_enrollments"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Enrollments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/ribbit/enrollments/status/course/{courseId}": {
			"get": {
				"operationId": "FindEnrollmentStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"enrollment_date": {
											"type": "string",
											"format": "date-time"
										},
										"progress": {
											"type": "number",
											"format": "double"
										},
										"is_enrolled": {
											"type": "boolean"
										}
									},
									"required": [
										"enrollment_date",
										"progress",
										"is_enrolled"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Enrollments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ribbit/enrollments/course/{courseId}": {
			"delete": {
				"operationId": "UnenrollStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Matrícula não encontrada"
					}
				},
				"tags": [
					"Enrollments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "courseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/ribbit/classes/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseClassGetResponse"
								}
							}
						}
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"CourseClass"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCourseClass",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseClassUpdateResponse"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao atualizar a aula"
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"CourseClass"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CourseClassUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCourseClass",
				"responses": {
					"204": {
						"description": "Aula deletada com sucesso"
					},
					"404": {
						"description": "Aula não encontrada"
					}
				},
				"tags": [
					"CourseClass"
				],
				"security": [
					{
						"bearerAuth": [
							"prof",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}